Introduce configuration file for MIDI profiles.
This configuration file will provide a way to map MIDI events to
mnemonics. This will be an opinionated mapping (i.e: there could very
different ways to define those mappings, not everyone wants the kick
to be on track 1), let's not try to be too flexible on this for now.