Rename midi_device to midi.
This file will likely contain a lot of short-lived classes until we
get a stable implementation, instead of creating a new file each time
and spending 25 seconds context-switching on this new file, let's have
a single one for now. We can split it later once the cost of scrolling
is too painful.