Implement some basic MIDI filtering.
This CL makes it possible to filter incoming MIDI messages, it
highlights that binary filters may not be the best approach here
considering some MIDI events like note off can have multiple possible
representations.

One example of this: note off can be represented as a dedicated status
message OR a note on event with a velocity of 0 (source: MIDI's
specification).