Introduce Mod class to represent a graphic unit.
Goal here is to have something that act a bit like an audio unit but
to draw stuff on a texture instead. Mods will be grouped in layers,
each layer will be a chain of mods, drawing to this layer. All layers
will then be merged in the main window.